+++ experiments/rbkit_command_test.rb
@@ -10,7 +10,11 @@ commands = [
   'stop_memory_profile',
   'objectspace_snapshot',
   'trigger_gc',
-  'handshake'
+  'handshake',
+  'use_cpu_time',
+  'use_wall_time',
+  'start_cpu_profiling',
+  'stop_cpu_profiling'
 ]
 
 output_file = File.open("/tmp/rbkit.log", "w")
@@ -20,10 +24,12 @@ ctx = ZMQ::Context.new
 puts "Enter IPv4 address of Rbkit server. (Blank for localhost) :"
 server_ip = gets.strip
 server_ip = "127.0.0.1" if server_ip.empty?
+pub_port = (ENV['RBKIT_PUB_PORT'] || 5555).to_i
+req_port = (ENV['RBKIT_REQ_PORT'] || 5556).to_i
 
 Thread.new do
   request_socket = ctx.socket(:REQ)
-  request_socket.connect("tcp://#{server_ip}:5556")
+  request_socket.connect("tcp://#{server_ip}:#{req_port}")
   loop do
     puts "Available commands :"
     commands.each_with_index do |c, i|
@@ -42,7 +48,7 @@ end
 
 socket = ctx.socket(:SUB)
 socket.subscribe("")
-socket.connect("tcp://#{server_ip}:5555")
+socket.connect("tcp://#{server_ip}:#{pub_port}")
 
 begin
   loop do