+++ ext/rbkit_time_helper.h
@@ -1,0 +1,8 @@
+#ifndef RBKIT_TIME_HELPER_H
+#define RBKIT_TIME_HELPER_H
+
+double get_cpu_time_in_usec();
+double get_wall_time_in_usec();
+double get_wall_time_in_msec();
+
+#endif