+++ spec/obj_created_spec.rb
@@ -1,6 +1,4 @@
 require 'spec_helper'
-require 'support/have_message_matcher'
-require 'msgpack'
 
 describe "obj_created event" do
   let(:payload) { Rbkit::MESSAGE_FIELDS[:payload] }
@@ -23,7 +21,7 @@ describe "obj_created event" do
         x[payload][class_name] =='ShortLivedBar' }
   end
   before(:all) do
-    server = Rbkit.start_profiling(enable_gc_stats: false, enable_object_trace: true)
+    server = Rbkit.start_profiling(enable_gc_stats: false, enable_object_trace: true, enable_cpu_profiling: false)
     @foo_obj = Foo.new
     packed_message = server.get_queued_messages
     Rbkit.stop_server