+++ CHANGELOG.md
@@ -21,3 +21,7 @@ WIP
 * Send server and protocol versions along with handshake response [#103](https://github.com/code-mancers/rbkit/pull/103)
 * Split object dump into smaller messages [#77](https://github.com/code-mancers/rbkit/pull/77)
 * Fix bug with handshake broken because of Rbkit::VERSION not being available.
+* Make REQ and PUB ports configurable using ENV variables
+* Always bind to 127.0.0.1
+* Add CPU Sampling profiler and cpu_sample message type. Bump up protocol version to v2.1
+* Use C89 std for compiling rbkit